Refactored dagger
