+++ app/src/main/kotlin/com/konifar/materialcat/di/ActivityComponent.kt
@@ -1,11 +1,14 @@
-package com.konifar.materialcat.presentation.di
+package com.konifar.materialcat.di
 
 import com.konifar.materialcat.MainActivity
 import com.konifar.materialcat.PhotoDetailActivity
 import dagger.Subcomponent
 
 @ActivityScope
-@Subcomponent(modules = arrayOf(ActivityModule::class))
+@Subcomponent(modules = arrayOf(
+        ActivityModule::class,
+        UseCaseModule::class
+))
 interface ActivityComponent {
 
     fun inject(activity: MainActivity)
