+++ app/src/main/kotlin/com/konifar/materialcat/di/FragmentComponent.kt
@@ -1,10 +1,13 @@
-package com.konifar.materialcat.presentation.di
+package com.konifar.materialcat.di
 
 import com.konifar.materialcat.presentation.gallery.GalleryFragment
 import dagger.Subcomponent
 
-@FragmentScope
-@Subcomponent(modules = arrayOf(FragmentModule::class))
+@ActivityScope
+@Subcomponent(modules = arrayOf(
+        FragmentModule::class,
+        UseCaseModule::class
+))
 interface FragmentComponent {
 
     fun inject(fragment: GalleryFragment)
