+++ app/src/main/kotlin/com/konifar/materialcat/extension/ContextExtensions.kt
@@ -1,7 +1,7 @@
 package com.konifar.materialcat.extension
 
 import android.content.Context
-import com.konifar.materialcat.ApplicationComponent
+import com.konifar.materialcat.di.ApplicationComponent
 import com.konifar.materialcat.MainApplication
 
 val Context.application: MainApplication
