+++ app/src/main/kotlin/com/konifar/materialcat/extension/FragmentExtensions.kt
@@ -1,8 +1,10 @@
 package com.konifar.materialcat.extension
 
 import android.support.v4.app.Fragment
-import com.konifar.materialcat.presentation.di.FragmentComponent
-import com.konifar.materialcat.presentation.di.FragmentModule
+import android.support.v7.app.AppCompatActivity
+import com.konifar.materialcat.di.FragmentComponent
+import com.konifar.materialcat.di.FragmentModule
+import com.konifar.materialcat.di.UseCaseModule
 
 val Fragment.component: FragmentComponent
-    get() = activity.component.plus(FragmentModule(this))
+    get() = (activity as AppCompatActivity).application.component.plus(FragmentModule(), UseCaseModule())
