+++ app/src/main/kotlin/com/konifar/materialcat/presentation/UseCase.kt
@@ -1,10 +1,6 @@
 package com.konifar.materialcat.presentation
 
-import org.greenrobot.eventbus.EventBus
-
 interface UseCase {
-
-    fun eventBus(): EventBus
 
     fun destroy()
