--- app/src/main/kotlin/com/konifar/materialcat/presentation/di/ActivityModule.kt
@@ -1,13 +1,0 @@
-package com.konifar.materialcat.presentation.di
-
-import android.support.v7.app.AppCompatActivity
-import dagger.Module
-import dagger.Provides
-
-@Module
-class ActivityModule(private val activity: AppCompatActivity) {
-
-    @Provides
-    fun provideActivity(): AppCompatActivity = activity
-
-}
