--- app/src/main/kotlin/com/konifar/materialcat/presentation/di/FragmentScope.kt
@@ -1,7 +1,0 @@
-package com.konifar.materialcat.presentation.di
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.RUNTIME)
-internal annotation class FragmentScope
